Move utils.h to utils folder
Also added function signatures for file operations
