+++ src/log_viewer/interface.c
@@ -1,7 +1,7 @@
 #include "log_viewer/interface.h"
 #include "windows/help_window.h"
 #include "windows/window_builder.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <string.h>
 #include <time.h>
