+++ src/log_viewer/main.c
@@ -2,7 +2,7 @@
 #include "tab_manager/tab_manager.h"
 #include "windows/window_builder.h"
 #include "context/context.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <ncurses.h>
 #include <string.h>
