+++ src/tab_manager/tab.c
@@ -1,5 +1,5 @@
 #include "tab_manager/tab.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <stdlib.h>
