+++ src/windows/help_window.c
@@ -1,5 +1,5 @@
 #include "windows/help_window.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <stdlib.h>
 #include <string.h>
