+++ src/windows/window_builder.c
@@ -1,5 +1,5 @@
 #include "windows/window_builder.h"
-#include "utils.h"
+#include "utils/utils.h"
 
 #include <string.h>
